refactoring model code

See merge request compsci308_2020spring/parser_parser_team21!34

+++ src/slogo/model/code/Token.java
@@ -2,5 +2,5 @@ package slogo.model.code;
 
 public interface Token {
     String toString();
-    int generateValue();
+    double generateValue();
 }
